Refine GUI design and enhance font engine
- Implemented PSF1/PSF2 font support with bold/italic simulation and text alignment.
- Added alpha blending and anti-aliasing (Wu's algorithm for lines, distance-based for shapes) to the software renderer.
- Introduced the "ACOS Synthetic Interface" design system with glassmorphism and neon accents.
- Updated GUI widgets (Button, TextBox, ListView, WindowWidget) and shell components (Taskbar, Launcher, NotificationCenter) to the new design system.
- Added acos::Span to the runtime library.
- Updated compositor to support alpha-blended windows.
- All changes comply with freestanding kernel requirements and AGENTS.md.

+++ kernel/graphics/types.h
@@ -3,7 +3,6 @@
 
 namespace acos::graphics {
 
-// Color utilities
 struct Color {
     u8 r, g, b, a;
     
@@ -15,18 +14,16 @@ struct Color {
     }
     
     static Color from_argb(u32 argb) {
-        return Color((argb >> 16) & 0xFF, (argb >> 8) & 0xFF, argb & 0xFF, (argb >> 24) & 0xFF);
+        return Color((u8)((argb >> 16) & 0xFF), (u8)((argb >> 8) & 0xFF), (u8)(argb & 0xFF), (u8)((argb >> 24) & 0xFF));
     }
 };
 
-// Gradient support
 struct Gradient {
     Color start;
     Color end;
     bool horizontal;
 };
 
-// Clipping rectangle
 struct ClipRect {
     i32 x, y, w, h;
